requirements.txt
dj-database-url==0.2.1
django-discover-runner==0.3
django-pipeline==1.3.3
docutils==0.10
Markdown==2.2.1
psycopg2==2.4.5

pytz
freezegun
hg+ssh://hg@bitbucket.org/fcurella/django-timedelta-field

# Django 1.5c2. Replace with Django==1.5 when 1.5 lands on pypi.
https://www.djangoproject.com/download/1.5c2/tarball/

# django-markupfield with Py3 support. Can be removed when 1.1 is released.
git+git://github.com/jamesturk/django-markupfield.git@70dc74

# My fork of django-sitetree, adding Py3 support. Can be removed when
# https://github.com/idlesign/django-sitetree/pull/93
git+git://github.com/jefftriplett/django-sitetree.git

# Aymeric's branch of South with Python 3 support. Can be removed when this
# is merged into master.
https://bitbucket.org/aaugustin/south-python3/get/a23d01c06065.zip
